Squash numerous compiler warnings
- add explicit casts where appropriate
- squash warnings about uninitialised/unused variables
- squash warnings relating to extern/static struct definitions

+++ src/shims/hw_config.h
@@ -54,7 +54,7 @@ _dispatch_get_logicalcpu_max(void)
 	(void)dispatch_assume(valsz == sizeof(uint32_t));
 #elif HAVE_SYSCONF && defined(_SC_NPROCESSORS_ONLN)
 	int ret = (int)sysconf(_SC_NPROCESSORS_ONLN);
-	val = ret < 0 ? 1 : ret;
+	val = ret < 0 ? 1 : (uint32_t)ret;
 #else
 #warning "no supported way to query logical CPU count"
 #endif
@@ -76,7 +76,7 @@ _dispatch_get_physicalcpu_max(void)
 	(void)dispatch_assume(valsz == sizeof(uint32_t));
 #elif HAVE_SYSCONF && defined(_SC_NPROCESSORS_ONLN)
 	int ret = (int)sysconf(_SC_NPROCESSORS_ONLN);
-	val = ret < 0 ? 1 : ret;
+	val = ret < 0 ? 1 : (uint32_t)ret;
 #else
 #warning "no supported way to query physical CPU count"
 #endif
@@ -98,7 +98,7 @@ _dispatch_get_activecpu(void)
 	(void)dispatch_assume(valsz == sizeof(uint32_t));
 #elif HAVE_SYSCONF && defined(_SC_NPROCESSORS_ONLN)
 	int ret = (int)sysconf(_SC_NPROCESSORS_ONLN);
-	val = ret < 0 ? 1 : ret;
+	val = ret < 0 ? 1 : (uint32_t)ret;
 #else
 #warning "no supported way to query active CPU count"
 #endif
